+++ lib/db/service/index.d.ts
@@ -1,3 +1,2 @@
-export * from './publish-model-change';
 export * from './sequelize-base.service';
 export * from './sequelize-base-service.interface';