+++ src/app/server/rfi-pub-sub-engine.ts
@@ -1,24 +1,41 @@
 import * as P from 'bluebird';
-import { Model } from 'sequelize-typescript';
+import { Model, Sequelize } from 'sequelize-typescript';
 import { GooglePubSub } from '@axelspringer/graphql-google-pubsub';
-import { status } from '@grpc/grpc-js';
 import { hostname } from 'os';
-import { NotificationOf, RfiSubscriptionOptions, uniqueSubscriptionNamePart } from '../../gql';
-import { publishModelChange } from '../../db/service/publish-model-change';
+import { RumbleshipBeeline } from '@rumbleship/o11y';
+import { ClassType } from './../../helpers/classtype';
+import {
+  ModelDelta,
+  NotificationOf,
+  RfiSubscriptionOptions,
+  uniqueSubscriptionNamePart,
+  NODE_CHANGE_NOTIFICATION
+} from '../../gql';
 import { RfiPubSubConfig } from '../config';
-import { DbModelAndOidScope } from './init-sequelize';
-import { RfiPubSubEngine } from './rfi-pub-sub-engine.interface';
+import { DbModelAndOidScope, getOidFor, getScopeFor } from './init-sequelize';
+import { RfiPubSubEngine, NodeChangePayload } from './rfi-pub-sub-engine.interface';
 
+import { CreateOptions, UpdateOptions, Model as SequelizeModel } from 'sequelize';
+import { getContextId } from '../rumbleship-context';
+/**
+ * @NOTE THIS IS IS ONLY FOR CLIENT SUBSCRIPTIONS
+ */
 export class RfiPubSub extends GooglePubSub implements RfiPubSubEngine {
   protected topicPrefix: string;
   public publisher_version: string;
   protected subscription_ids: number[];
-  constructor(publisher_version: string, config: RfiPubSubConfig) {
+  protected beeline_cls: ClassType<RumbleshipBeeline> & typeof RumbleshipBeeline;
+  constructor(
+    publisher_version: string,
+    config: RfiPubSubConfig,
+    beeline: ClassType<RumbleshipBeeline> & typeof RumbleshipBeeline
+  ) {
     RfiPubSub.validatePubSubConfig(config);
     const { topicPrefix, keyFilename } = config;
     super(keyFilename === `/dev/null` ? {} : config, uniqueSubscriptionNamePart);
     this.topicPrefix = topicPrefix;
     this.publisher_version = publisher_version;
+    this.beeline_cls = beeline;
     this.subscription_ids = [];
   }
 
@@ -39,7 +56,45 @@ export class RfiPubSub extends GooglePubSub implements RfiPubSubEngine {
     }
   }
 
-  // Couldn't get typescript to be happy with 'extends', so we end up repeat ourselves
+  /**
+   *
+   * @param {Sequelize } sequelize
+   *
+   * @description Attaches global model hooks, respecting transactions, to th
+   */
+  public linkToSequelize(sequelize: Sequelize) {
+    const hookCb = (pubSub: RfiPubSub, notification_of: NotificationOf) => {
+      return function publisherHook(
+        sequelize_instance: SequelizeModel<any, any>,
+        options: CreateOptions | UpdateOptions
+      ) {
+        // A slight incompatibility between types defined in `sequelize-typescript` and parent `sequelize`:
+        // definition of the instance passed to the hooks, so we cast to generic Model<any,any>
+        const instance = sequelize_instance as Model<any, any>;
+        // Cache the delta of change in closure so we can publish original values __after__ the
+        // commit, which overwrites those original values.
+        const deltas = getChangedAttributes(instance);
+
+        if (options && options.transaction) {
+          const context_id = getContextId(options.transaction);
+          options.transaction.afterCommit(t => {
+            pubSub.publishModelChange(notification_of, instance, deltas, context_id);
+          });
+        } else {
+          pubSub.publishModelChange(notification_of, instance, deltas);
+        }
+      };
+    };
+
+    sequelize.afterCreate(hookCb(this, NotificationOf.CREATED));
+    sequelize.afterUpdate(hookCb(this, NotificationOf.UPDATED));
+    // sequelize.afterBulkCreate((instances, options) => gqlBulkCreateHook(pubSub, instances, options));
+  }
+
+  public getMarshalledTraceContext(context_id: string): string {
+    return this.beeline_cls.marshalTraceContext(this.beeline_cls.getTraceContext(context_id));
+  }
+
   public async publish(triggerName: string, payload: any): Promise<void> {
     const topicName = `${this.topicPrefix}_${triggerName}`;
     await this.createTopicIfNotExist(topicName);
@@ -70,13 +125,39 @@ export class RfiPubSub extends GooglePubSub implements RfiPubSubEngine {
     }
     this.subscription_ids = [];
   }
-  public asyncIterator<T>(triggers: string | string[]): AsyncIterator<T> {
-    return super.asyncIterator(triggers);
-  }
 
-  public publishPayload(notificationType: NotificationOf, model: Model, deltas: any[]): void {
+  /**
+   *
+   * @param notification
+   * @param model
+   * @param deltas
+   *
+   * @note This triggers floating promises which is explicity does not await!
+   */
+  publishModelChange(
+    notification: NotificationOf,
+    model: Model,
+    deltas: ModelDelta[],
+    context_id?: string
+  ): void {
+    const rval = payloadFromModel(model) as NodeChangePayload;
+    rval.action = notification;
+    rval.deltas = deltas;
+    rval.publisher_version = this.publisher_version;
+    rval.marshalled_trace = context_id ? this.getMarshalledTraceContext(context_id) : undefined;
+
+    const payload = JSON.stringify(rval);
+
+    const oidScope = getScopeFor(model);
+    const topicName: string = `${NODE_CHANGE_NOTIFICATION}_${oidScope}`;
+
+    // Publish the change on most generic topic
     // tslint:disable-next-line: no-floating-promises
-    publishModelChange(this, notificationType, model, deltas);
+    this.publish(NODE_CHANGE_NOTIFICATION, payload);
+
+    // Also publish the change topic specific to _this_ model
+    // tslint:disable-next-line: no-floating-promises
+    this.publish(topicName, payload);
   }
 
   public async deleteCurrentSubscriptionsMatchingPrefix() {
@@ -99,18 +180,43 @@ export class RfiPubSub extends GooglePubSub implements RfiPubSubEngine {
   }
 
   private async createTopicIfNotExist(topicName: string): Promise<void> {
+    const TOPIC_ALREADY_EXISTS_GCP_MAGIC_NUMBER = 6;
     const topic = this.pubSubClient.topic(topicName);
     const [exists] = await topic.exists();
     if (!exists) {
       try {
         await this.pubSubClient.createTopic(topicName);
       } catch (e) {
-        if (!(e.code === status.ALREADY_EXISTS)) {
-          // It can be created during a race condition,
-          // so only rethrow if it is another error
+        if (!(e.code === TOPIC_ALREADY_EXISTS_GCP_MAGIC_NUMBER)) {
+          // A topic can be created many times concurrently; Google only lets one get created
+          // and throws a specific error for those that fail. Only rethrow.
           throw e;
         }
       }
     }
   }
+}
+
+function payloadFromModel(model: Model): { oid: string; id: string } {
+  const modelId = model?.get('id') as string;
+  const oid = getOidFor(model).toString();
+  return { oid, id: modelId };
+}
+
+function getChangedAttributes(instance: Model<any, any>): ModelDelta[] {
+  const deltas: ModelDelta[] = [];
+  const values = instance.get({ plain: true });
+  for (const key in values) {
+    if (values.hasOwnProperty(key)) {
+      if (instance.changed(key as any)) {
+        const delta: ModelDelta = {
+          key,
+          previousValue: instance.previous(key as any),
+          newValue: instance.get(key as any)
+        };
+        deltas.push(delta);
+      }
+    }
+  }
+  return deltas;
 }