+++ src/gql/relay/relay.interface.ts
@@ -78,7 +78,6 @@ export interface NodeService<T> {
   getContext(): RumbleshipContext;
   getServiceFor<S extends Node<S>, V extends NodeService<S>>(cls: ClassType<S> | string): V;
   setServiceRegister(services: any): void;
-  publishLastKnownState(oid: Oid): Promise<void>;
   newTransaction(params: {
     isolation: NodeServiceIsolationLevel;
     autocommit: boolean;