+++ CHANGELOG.md
@@ -8,8 +8,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [UNRELEASED]
 
 ### Added
+  * Propagate trace context over the pub/sub bus
+  * Generic traceContext added for `node.id`, various filters inside `SequelizeBaseService`
+  * Definition of the `SubscriptionCommandPayload`, inherits from generic `Payload`
 ### Removed
+  * dependency on `@google/grpc-js`
+  * SpyglassLogger interface from exports
 ### Changed
+  * Peer dependency is `@rumbleship/o11y#4.0.1`
+  * Subscription-related interface called `Payload` now called `NodeChangePayload`, inherits from generic `Payload`
 ### Fixed
 ### Deprecated
 ### Security